+++ workers/unity/Assets/Playground/Scripts/Cubes/CollisionProcessSystem.cs
@@ -26,10 +26,9 @@ namespace Playground
 
             collisionGroup = GetEntityQuery(
                 ComponentType.ReadOnly<Launchable.Component>(),
-                ComponentType.ReadOnly<Launchable.ComponentAuthority>(),
+                ComponentType.ReadOnly<Launchable.HasAuthority>(),
                 ComponentType.ReadOnly<CollisionComponent>()
             );
-            collisionGroup.SetFilter(Launchable.ComponentAuthority.Authoritative);
         }
 
         protected override void OnUpdate()