+++ workers/unity/Assets/Playground/Scripts/Player/MoveLocalPlayerSystem.cs
@@ -34,18 +34,16 @@ namespace Playground
 
             newPlayerGroup = GetEntityQuery(
                 ComponentType.ReadOnly<PlayerInput.Component>(),
-                ComponentType.ReadOnly<PlayerInput.ComponentAuthority>(),
+                ComponentType.ReadOnly<PlayerInput.HasAuthority>(),
                 ComponentType.Exclude<Speed>()
             );
-            newPlayerGroup.SetFilter(PlayerInput.ComponentAuthority.Authoritative);
 
             playerInputGroup = GetEntityQuery(
                 ComponentType.ReadWrite<Rigidbody>(),
                 ComponentType.ReadWrite<Speed>(),
                 ComponentType.ReadOnly<PlayerInput.Component>(),
-                ComponentType.ReadOnly<TransformInternal.ComponentAuthority>()
+                ComponentType.ReadOnly<TransformInternal.HasAuthority>()
             );
-            playerInputGroup.SetFilter(TransformInternal.ComponentAuthority.Authoritative);
         }
 
         protected override void OnUpdate()
@@ -84,7 +82,7 @@ namespace Playground
                     var speedSmoothVelocity = speed.SpeedSmoothVelocity;
 
                     currentSpeed = Mathf.SmoothDamp(currentSpeed, targetSpeed, ref speedSmoothVelocity, SpeedSmoothTime,
-                        MaxSpeed, Time.deltaTime);
+                        MaxSpeed, Time.DeltaTime);
                     speed = new Speed
                     {
                         CurrentSpeed = currentSpeed,