+++ workers/unity/Packages/io.improbable.gdk.core/NetworkStats/NetworkStatisticsSystem.cs
@@ -50,7 +50,7 @@ namespace Improbable.Gdk.Core.NetworkStats
 
             lastIncomingData.Clear();
             lastOutgoingData.Clear();
-            lastFrameTime = Time.deltaTime;
+            lastFrameTime = Time.DeltaTime;
         }
 
         [Conditional("UNITY_EDITOR")]