+++ workers/unity/Packages/io.improbable.gdk.core/Worker/WorkerInWorld.cs
@@ -67,6 +67,7 @@ namespace Improbable.Gdk.Core
         private void AddCoreSystems()
         {
             World.CreateSystem<WorkerSystem>(this);
+            World.GetOrCreateSystem<UpdateWorldTimeSystem>();
             World.GetOrCreateSystem<CommandSystem>();
             World.GetOrCreateSystem<ComponentUpdateSystem>();
             World.GetOrCreateSystem<EntitySystem>();