+++ workers/unity/Packages/io.improbable.gdk.playerlifecycle/Systems/PlayerHeartbeat/SendPlayerHeartbeatRequestSystem.cs
@@ -9,7 +9,7 @@ namespace Improbable.Gdk.PlayerLifecycle
     [UpdateBefore(typeof(SpatialOSUpdateGroup))]
     public class SendPlayerHeartbeatRequestSystem : ComponentSystem
     {
-        private float timeOfNextHeartbeat = Time.time + PlayerLifecycleConfig.PlayerHeartbeatIntervalSeconds;
+        private double timeOfNextHeartbeat;
         private EntityQuery group;
         private CommandSystem commandSystem;
 
@@ -17,24 +17,25 @@ namespace Improbable.Gdk.PlayerLifecycle
         {
             base.OnCreate();
 
+            timeOfNextHeartbeat = Time.ElapsedTime + PlayerLifecycleConfig.PlayerHeartbeatIntervalSeconds;
+
             group = GetEntityQuery(
-                ComponentType.ReadOnly<PlayerHeartbeatServer.ComponentAuthority>(),
+                ComponentType.ReadOnly<PlayerHeartbeatServer.HasAuthority>(),
                 ComponentType.ReadWrite<HeartbeatData>(),
                 ComponentType.ReadOnly<SpatialEntityId>()
             );
-            group.SetFilter(PlayerHeartbeatServer.ComponentAuthority.Authoritative);
 
             commandSystem = World.GetExistingSystem<CommandSystem>();
         }
 
         protected override void OnUpdate()
         {
-            if (Time.time < timeOfNextHeartbeat)
+            if (Time.ElapsedTime < timeOfNextHeartbeat)
             {
                 return;
             }
 
-            timeOfNextHeartbeat = Time.time + PlayerLifecycleConfig.PlayerHeartbeatIntervalSeconds;
+            timeOfNextHeartbeat = Time.ElapsedTime + PlayerLifecycleConfig.PlayerHeartbeatIntervalSeconds;
 
             Entities.With(group).ForEach((ref SpatialEntityId spatialEntityId) =>
             {