+++ workers/unity/Packages/io.improbable.gdk.transformsynchronization/Systems/GetTransformValueToSetSystem.cs
@@ -17,12 +17,18 @@ namespace Improbable.Gdk.TransformSynchronization
 
             worker = World.GetExistingSystem<WorkerSystem>();
 
-            transformGroup = GetEntityQuery(
-                ComponentType.ReadWrite<BufferedTransform>(),
-                ComponentType.ReadWrite<TransformToSet>(),
-                ComponentType.ReadOnly<TransformInternal.ComponentAuthority>()
-            );
-            transformGroup.SetFilter(TransformInternal.ComponentAuthority.NotAuthoritative);
+            transformGroup = GetEntityQuery(new EntityQueryDesc
+            {
+                All = new[]
+                {
+                    ComponentType.ReadWrite<BufferedTransform>(),
+                    ComponentType.ReadWrite<TransformToSet>(),
+                },
+                None = new[]
+                {
+                    ComponentType.ReadOnly<TransformInternal.HasAuthority>(),
+                }
+            });
         }
 
         protected override void OnUpdate()