+++ workers/unity/Packages/io.improbable.gdk.transformsynchronization/Systems/InterpolateTransformSystem.cs
@@ -18,15 +18,21 @@ namespace Improbable.Gdk.TransformSynchronization
 
             updateSystem = World.GetExistingSystem<ComponentUpdateSystem>();
 
-            interpolationGroup = GetEntityQuery(
-                ComponentType.ReadWrite<BufferedTransform>(),
-                ComponentType.ReadWrite<DeferredUpdateTransform>(),
-                ComponentType.ReadOnly<TransformInternal.Component>(),
-                ComponentType.ReadOnly<SpatialEntityId>(),
-                ComponentType.ReadOnly<InterpolationConfig>(),
-                ComponentType.ReadOnly<TransformInternal.ComponentAuthority>()
-            );
-            interpolationGroup.SetFilter(TransformInternal.ComponentAuthority.NotAuthoritative);
+            interpolationGroup = GetEntityQuery(new EntityQueryDesc
+            {
+                All = new[]
+                {
+                    ComponentType.ReadWrite<BufferedTransform>(),
+                    ComponentType.ReadWrite<DeferredUpdateTransform>(),
+                    ComponentType.ReadOnly<TransformInternal.Component>(),
+                    ComponentType.ReadOnly<SpatialEntityId>(),
+                    ComponentType.ReadOnly<InterpolationConfig>(),
+                },
+                None = new[]
+                {
+                    ComponentType.ReadOnly<TransformInternal.HasAuthority>()
+                }
+            });
         }
 
         protected override void OnUpdate()